+++ src/server/game/DungeonFinding/LFGMgr.cpp
@@ -154,10 +154,10 @@ void LFGMgr::LoadRewards()
         uint32 dungeonId = fields[0].GetUInt32();
         uint32 maxLevel = fields[1].GetUInt8();
         uint32 firstQuestId = fields[2].GetUInt32();
-        uint32 firstMoneyVar = fields[3].GetUInt32();
+        uint64 firstMoneyVar = fields[3].GetUInt64();
         uint32 firstXPVar = fields[4].GetUInt32();
         uint32 otherQuestId = fields[5].GetUInt32();
-        uint32 otherMoneyVar = fields[6].GetUInt32();
+        uint64 otherMoneyVar = fields[6].GetUInt64();
         uint32 otherXPVar = fields[7].GetUInt32();
 
         if (!sLFGDungeonStore.LookupEntry(dungeonId))
