+++ src/server/game/Entities/Pet/Pet.cpp
@@ -1710,7 +1710,7 @@ bool Pet::ResetTalents(bool no_cost)
         return false;
     }
 
-    uint32 cost = 0;
+    uint64 cost = 0;
 
     if (!no_cost)
     {
@@ -1768,7 +1768,7 @@ bool Pet::ResetTalents(bool no_cost)
 
     if (!no_cost)
     {
-        player->ModifyMoney(-(int32)cost);
+        player->ModifyMoney(-(int64)cost);
 
         m_resetTalentsCost = cost;
         m_resetTalentsTime = time(NULL);
